Public Version of Web Spider
Crawls a given domain (i.e. www.google.com or www.apple.com/store) and
counts links to external sites (i.e. any link that is not contained on
the given domain crawl). Works well with HTML, not 100% reliable
however. Use caution when running on large sites as this program does
download each file on the domain and it can take significant time,
depending on the server and network connection. This version is also
only single threaded, therefore it is more limited as well.